requirements.txt
# Core dependencies
boto3>=1.26.0
watchdog>=3.0.0
PyYAML>=6.0
python-dotenv>=1.0.0

# Async HTTP server for HLS
aiohttp>=3.8.0

# Database
# sqlite3 is built into Python

# Utilities
click>=8.1.0          # CLI interface

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
